refactor(web): deduplicate shared ui primitives
