+++ web/src/components/index.ts
@@ -16,6 +16,9 @@ export * from './CountersContext';
 export * from './ConfigTabs';
 export * from './ConfigTabStrip';
 export * from './SaveDiffModal';
+export * from './SideBySideDiff';
+export * from './MetricSparkline';
+export * from './SearchInput';
 export { default as YamlIOModal } from './YamlIOModal';
 export type { YamlIOModalProps, YamlIOMode } from './YamlIOModal';
 export { default as BulkBar } from './BulkBar';
