+++ web/src/pages/builtin/devices/DevicePageHeader.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { Button, Flex, Icon, Text, TextInput } from '@gravity-ui/uikit';
-import { Magnifier, Plus } from '@gravity-ui/icons';
+import { Button, Flex, Icon, Text } from '@gravity-ui/uikit';
+import { Plus } from '@gravity-ui/icons';
+import { SearchInput } from '../../../components';
 
 export interface DevicePageHeaderProps {
     onCreateDevice: () => void;
@@ -21,18 +22,11 @@ export const DevicePageHeader: React.FC<DevicePageHeaderProps> = ({
             <Text variant="header-1">Devices</Text>
             <Flex grow />
             <div style={{ flexBasis: 380, flexShrink: 1 }}>
-                <TextInput
+                <SearchInput
                     controlRef={searchRef}
                     value={searchQuery}
                     onUpdate={onSearchChange}
                     placeholder="Search devices… (⌘K)"
-                    startContent={
-                        <Flex alignItems="center" justifyContent="center" style={{ paddingInline: 8, color: 'var(--g-color-text-hint)' }}>
-                            <Icon data={Magnifier} size={16} />
-                        </Flex>
-                    }
-                    hasClear
-                    type="search"
                 />
             </div>
             <Button view="action" onClick={onCreateDevice}>
