+++ web/src/pages/builtin/pipelines/PipelinesPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useCallback, useRef, useMemo } from 'react';
-import { Button, Flex, Icon, Text, TextInput } from '@gravity-ui/uikit';
-import { Magnifier, Plus } from '@gravity-ui/icons';
-import { PageLayout, PageLoader, EmptyState } from '../../../components';
+import { Button, Flex, Icon, Text } from '@gravity-ui/uikit';
+import { Plus } from '@gravity-ui/icons';
+import { PageLayout, PageLoader, EmptyState, SearchInput } from '../../../components';
 import { usePipelinesData } from './hooks/usePipelinesData';
 import { useUnsavedChangesBlocker } from '../_shared/lane-editor';
 import { PipelineCard } from './components/PipelineCard';
@@ -66,18 +66,11 @@ const PipelinesPage = (): React.JSX.Element => {
             <Text variant="header-1">Pipelines</Text>
             <Flex grow />
             <div style={{ flexBasis: 380, flexShrink: 1 }}>
-                <TextInput
+                <SearchInput
                     controlRef={searchRef}
                     value={searchQuery}
                     onUpdate={setSearchQuery}
                     placeholder="Search pipelines, functions… (⌘K)"
-                    startContent={
-                        <Flex alignItems="center" justifyContent="center" style={{ paddingInline: 8, color: 'var(--g-color-text-hint)' }}>
-                            <Icon data={Magnifier} size={16} />
-                        </Flex>
-                    }
-                    hasClear
-                    type="search"
                 />
             </div>
             <Button
