+++ web/src/pages/modules/acl/Sparkline.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { MetricSparkline } from '../../../components';
 
 interface SparklineProps {
     /** Data points to render. Null means no data is available. */
@@ -9,10 +10,6 @@ interface SparklineProps {
     fill?: boolean;
 }
 
-/**
- * Pure SVG sparkline component for ACL NG counter cells.
- * When values is null or empty, renders an empty placeholder slot.
- */
 const Sparkline: React.FC<SparklineProps> = ({
     values,
     width = 64,
@@ -20,60 +17,29 @@ const Sparkline: React.FC<SparklineProps> = ({
     color = 'var(--fw-accent)',
     fill = true,
 }) => {
-    if (!values || values.length < 2) {
-        return (
-            <span
-                className="fw-spark-empty"
-                title="No counter history available"
-                style={{
-                    display: 'inline-flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    width,
-                    height,
-                }}
-            >
-                --
-            </span>
-        );
-    }
-
-    const max = Math.max(1, ...values);
-    const min = Math.min(0, ...values);
-    const range = Math.max(1, max - min);
-    const step = width / (values.length - 1 || 1);
-
-    const points = values.map((v, idx) => {
-        const x = idx * step;
-        const y = height - ((v - min) / range) * (height - 2) - 1;
-        return [x, y] as [number, number];
-    });
-
-    const path = points
-        .map((p, idx) => `${idx === 0 ? 'M' : 'L'}${p[0].toFixed(1)},${p[1].toFixed(1)}`)
-        .join(' ');
-    const fillPath = `${path} L${width},${height} L0,${height} Z`;
-    const last = points[points.length - 1];
-
     return (
-        <svg
-            width={width}
-            height={height}
-            viewBox={`0 0 ${width} ${height}`}
-            className="fw-spark-svg"
-            aria-hidden="true"
-        >
-            {fill && <path d={fillPath} fill={color} opacity="0.16" />}
-            <path
-                d={path}
-                fill="none"
-                stroke={color}
-                strokeWidth="1.25"
-                strokeLinejoin="round"
-                strokeLinecap="round"
-            />
-            {last && <circle cx={last[0]} cy={last[1]} r="1.6" fill={color} />}
-        </svg>
+        <MetricSparkline values={values} width={width} height={height} color={color} fill={fill}>
+            {({ base, hasData }) => {
+                if (hasData) {
+                    return base;
+                }
+                return (
+                    <span
+                        className="fw-spark-empty"
+                        title="No counter history available"
+                        style={{
+                            display: 'inline-flex',
+                            alignItems: 'center',
+                            justifyContent: 'center',
+                            width,
+                            height,
+                        }}
+                    >
+                        --
+                    </span>
+                );
+            }}
+        </MetricSparkline>
     );
 };
