+++ web/src/pages/operators/route/RoutePage.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useMemo, useRef, useState } from 'react';
-import { Button, Flex, Icon, Text, TextInput } from '@gravity-ui/uikit';
-import { Magnifier, Plus } from '@gravity-ui/icons';
-import { PageLayout, PageLoader, ConfigTabStrip, BulkBar } from '../../../components';
+import { Button, Flex, Icon, Text } from '@gravity-ui/uikit';
+import { Plus } from '@gravity-ui/icons';
+import { PageLayout, PageLoader, ConfigTabStrip, BulkBar, SearchInput } from '../../../components';
 import { AddConfigModal, BulkDeleteModal } from '../../_shared/draft';
 import { API } from '../../../api';
 import { toaster } from '../../../utils';
@@ -220,18 +220,11 @@ const RoutePage: React.FC = () => {
             <Text variant="header-1">Routing Table</Text>
             <Flex grow />
             <div style={{ flexBasis: 380, flexShrink: 1 }}>
-                <TextInput
-                    controlRef={searchRef as React.RefObject<HTMLInputElement | null>}
+                <SearchInput
+                    controlRef={searchRef}
                     value={search}
                     onUpdate={setSearch}
                     placeholder="Search prefix, nexthop or peer… (/)"
-                    startContent={
-                        <Flex alignItems="center" justifyContent="center" style={{ paddingInline: 8, color: 'var(--g-color-text-hint)' }}>
-                            <Icon data={Magnifier} size={16} />
-                        </Flex>
-                    }
-                    hasClear
-                    type="search"
                 />
             </div>
             <Button view="outlined" onClick={handleFlush} disabled={!currentConfig}>
